Added app_frontend NOT WORKING (!!!).

+++ prerequisites.sh
@@ -4,13 +4,16 @@
 # node 18.12.1
 # npm 6.14.4
 
+# XXX put all to true
 do_apt_update="true";
 do_apt_upgrade="true";
 docker_clean_containers="true";
 docker_clean_images="true";
 docker_system_prune="true";
+docker_image_prune="true";
 reinstall_containerd_docker_ce="true";
 docker_stop="true";
+install_dependencies="true";
 
 do_docker_stop()
 {
@@ -24,9 +27,13 @@ do_docker_stop()
 	set -x;
 
 	[ "$docker_clean_containers" = "true" ] && \
+		[ "$(docker ps -aq)" != "" ] && \
 		docker container rm -f $(docker ps -aq) 2>/dev/null;
-	[ "$docker_clean_images" = "true" ] && docker rmi $(docker images -aq) 2>/dev/null;
-	[ "$docker_system_prune" = "true" ] && docker system prune -y 2>/dev/null;
+	[ "$docker_clean_images" = "true" ] && \
+		[ "$(docker images -aq)" != "" ] && \
+		docker rmi $(docker images -aq);
+	[ "$docker_system_prune" = "true" ] && docker system prune;
+	[ "$docker_image_prune" = "true" ] && docker image prune;
 	[ "$reinstall_containerd_docker_ce" = "true" ] && \
 		sudo apt install -y containerd.io docker-ce;
 	[ "$docker_stop" = "true" ] && do_docker_stop;
@@ -34,47 +41,53 @@ do_docker_stop()
 	[ "$do_apt_update" = "true" ] && sudo apt update;
 	[ "$do_apt_upgrade" = "true" ] && sudo apt -y upgrade && sudo apt -y autoremove;
 
-	sudo dpkg --configure -a
-	sudo apt install -y docker-compose-plugin;
-	sudo apt install -y npm;
-	sudo npm install -g n;
-	sudo n stable;
-	sudo npm install -g npm@latest;
-	sudo npm install -g @nestjs/cli@latest;
+	if [ "$install_dependencies" = "true" ] ; then
+
+		sudo dpkg --configure -a
+		sudo apt install -y docker-compose-plugin;
+		sudo apt install -y npm;
+		sudo npm install -g n;
+		sudo n stable;
+		sudo npm install -g npm@latest;
+		sudo npm install -g @nestjs/cli@latest;
 
 # For docker-compose in rootless mode:
-	sudo apt install -y uidmap; # updates even if already installed.
+		sudo apt install -y uidmap; # updates even if already installed.
 # Script to check user subordinates:
-	subuid=`grep ^$(whoami): /etc/subuid | awk -F ":" '{print $3}'`;
-	[ "$subuid" -lt "65536" ] && echo "Subordinate uid $subuid fail." && exit 1;
-	subgid=`grep ^$(whoami): /etc/subgid | awk -F ":" '{print $3}'`;
-	[ "$subgid" -lt "65536" ] && echo "Subordinate gid $subgid fail." && exit 2;
+		subuid=`grep ^$(whoami): /etc/subuid | awk -F ":" '{print $3}'`;
+		[ "$subuid" -lt "65536" ] && echo "Subordinate uid $subuid fail." && exit 1;
+		subgid=`grep ^$(whoami): /etc/subgid | awk -F ":" '{print $3}'`;
+		[ "$subgid" -lt "65536" ] && echo "Subordinate gid $subgid fail." && exit 2;
 # Needed by rootless:
-	sudo apt install -y dbus-user-session;
-	sudo apt install -y fuse-overlayfs; # recomended
+		sudo apt install -y dbus-user-session;
+		sudo apt install -y fuse-overlayfs; # recomended
 # Rootless docker requires version of slirp4netns greater than v0.4.0 \
 # (when vpnkit is not installed). Check you have this with
 #	slirp4netns --version
 # If you do not have this download and install with \
 #	sudo apt install -y slirp4netns
 # Script to check about this:
-	vp1=`slirp4netns --version | awk '{print $3}' | awk -F "." '{print $1}'`;
-	vp2=`slirp4netns --version | awk '{print $3}' | awk -F "." '{print $2}'`;
-	[ "$vp1" -gt "0" ] && need_install="false" ||
-		[ "$vp2" -ge "4" ] && need_install="false" ||
-			need_install="true";
-	[ "$need_install" = "true" ] && sudo apt install -y slirp4netns;
+		vp1=`slirp4netns --version | awk '{print $3}' | awk -F "." '{print $1}'`;
+		vp2=`slirp4netns --version | awk '{print $3}' | awk -F "." '{print $2}'`;
+		[ "$vp1" -gt "0" ] && need_install="false" ||
+			[ "$vp2" -ge "4" ] && need_install="false" ||
+				need_install="true";
+		[ "$need_install" = "true" ] && sudo apt install -y slirp4netns;
 
-	[ "$reinstall_containerd_docker_ce" = "true" ] && \
-		rm -f ~/.config/systemd/user/docker.service;
+		[ "$reinstall_containerd_docker_ce" = "true" ] && \
+			rm -f ~/.config/systemd/user/docker.service;
+
 # Run
 #	dockerd-rootless-setuptool.sh install
 # as a non-root user to set up the daemon. If this fails, run
 #	sudo apt install -y docker-ce-rootless-extras
 # Script to do so:
-	dockerd-rootless-setuptool.sh install 2>/dev/null || \
-		(sudo apt install -y docker-ce-rootless-extras && \
-			dockerd-rootless-setuptool.sh --force install);
+		dockerd-rootless-setuptool.sh install 2>/dev/null || \
+			(sudo apt install -y docker-ce-rootless-extras && \
+				dockerd-rootless-setuptool.sh --force install);
+
+	fi;
+
 # The systemd unit file is installed as ~/.config/systemd/user/docker.service.
 	systemctl --user start docker
 	docker context use rootless
